add books to the database
